**Before**

```
digest64 vs hash4Inputs vs hash8HashObjects
    ✔ digest64 50023 times                                                27.63268 ops/s    36.18903 ms/op   x0.957          7 runs  0.778 s
    ✔ hash 200092 times using hash4Inputs                                 4.879145 ops/s    204.9539 ms/op   x0.962          3 runs   1.24 s
    ✔ hash 200092 times using hash8HashObjects                            5.783024 ops/s    172.9199 ms/op   x0.995          3 runs   1.04 s
```

**After**

```
digest64 vs hash4Inputs vs hash8HashObjects
    ✔ digest64 50023 times                                                27.29671 ops/s    36.63445 ms/op   x0.962         10 runs  0.868 s
    ✔ hash 200092 times using hash4Inputs                                 8.757761 ops/s    114.1844 ms/op   x1.132          4 runs   1.04 s
    ✔ hash 200092 times using hash8HashObjects                            9.278248 ops/s    107.7790 ms/op   x1.207          3 runs  0.870 s
```